feat: split framework/CLI versioning with independent release workflows
Decouple framework and CLI release cycles so each component can evolve
independently. Framework uses `fw-*` tags, CLI uses `cli-*` tags, with
separate GitHub Actions workflows for each.

- Reset CLI version to 1.0.0, bump framework manifest to 2.1.0
- Add `get_latest_release_by_prefix()` to filter releases by tag prefix
- Add `strip_tag_prefix()` helper for fw-/cli-/v- prefixes
- Rename `update` command to `update-framework`, create combined `update`
- Add `update-framework` subcommand to CLI router
- Show both framework and CLI versions in `status` and `about`
- Split release.yml into release-framework.yml and release-cli.yml
- Update install scripts to fetch `cli-*` releases
